expr: drop the double must-use and a same-type cast from the time-value getter
get_time_value returned Result behind an explicit must_use attribute, which clippy reads as double-must-use because Result already carries it, and the get_time_value year read cast an i32 core field to i32. tidb-expr stays 1199/1199 with the documented network flake filtered.

+++ rust/crates/tidb-expr/src/lib.rs
@@ -486,7 +486,6 @@ pub fn is_valid_current_timestamp_expr(
 /// the small arithmetic form the source helper evaluates before parsing.
 /// Unknown AST forms preserve Go's zero-value datum (`NULL`) rather than
 /// pretending to evaluate a wider build-context surface.
-#[must_use]
 pub fn get_time_value(
     cols: &dyn Columns,
     expr: &Expr,
@@ -624,7 +623,7 @@ fn current_date_value(
     let current = current_time_value(cols, kind, fsp)?;
     let core = current.core_time();
     tidb_datatype::Time::from_date_checked(
-        core.year() as i32,
+        core.year(),
         core.month() as i32,
         core.day() as i32,
         0,
